flytter spm om selvstendig og frilans til steg 1
